% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_year_column.R
\name{add_year_column}
\alias{add_year_column}
\title{Adds the year column}
\usage{
add_year_column(list_of_datasets)
}
\arguments{
\item{list_of_datasets}{A list containing named datasets}
}
\value{
A list of datasets with the year column
}
\description{
This function works by extracting the year string contained in
the data set name and appending a new column to the data set with the numeric
value of the year. This means that the data sets have to have a name of the
form data_set_2001 or data_2001_europe, etc
}
\examples{
\dontrun{
#`list_of_data_sets` is a list containing named data sets
# For example, to access the first data set, called dataset_1 you would
# write
list_of_data_sets$dataset_1
add_year_column(list_of_data_sets)
}
}
