% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_data_to_upper.R
\name{all_data_to_upper}
\alias{all_data_to_upper}
\title{Converts column names to upper case for a given list of datasets}
\usage{
all_data_to_upper(list_of_loaded_datasets)
}
\arguments{
\item{list_of_loaded_datasets}{A list of loaded datasets}
}
\value{
Changes the datasets in place, with column names converted to upper case
}
\description{
Converts column names to upper case for a given list of datasets
}
\examples{
\donturn{
setwd("path/to/datasets/")
list_of_loaded_datasets <- ls(pattern = "expr") # "expr" is a character string common to
every data set (for example, suppose all you loaded data sets are of the form *_entrep_*)
updated_list <- all_data_to_upper(list_of_loaded_datasets)
}
}

