% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/around.R
\name{around}
\alias{around}
\title{Is x around y?}
\usage{
around(x, y, eps)
}
\arguments{
\item{x}{The value to test}

\item{y}{Value of interest}

\item{eps}{The threshold}
}
\value{
TRUE if x is in the interval (y - eps, y + eps)
}
\description{
Is x around y?
}
\examples{
around(2, 3, 0.5)
around(2, 3, 1)
}

