% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_id.R
\name{create_group_id}
\alias{create_group_id}
\title{Creates a group id. Especially useful in the case of panel data}
\usage{
create_group_id(list_of_loaded_datasets, list_of_variables)
}
\arguments{
\item{list_of_loaded_datasets}{A list of loaded datasets}

\item{list_of_variables}{A list of variables by which to group}
}
\value{
A data frame with the column `group_id` added to it
}
\description{
Creates a group id. Especially useful in the case of panel data
}
\examples{
\dontrun{
setwd("path/to/datasets/")
list_of_datasets <- ls(pattern = "expr") # "expr" is a character string common to
every data set (for example, suppose all you loaded data sets are of the form *_entrep_*)
list_of_loaded_datasets <- group_id(list_of_datasets)
updated_list <- all_data_to_upper(list_of_loaded_datasets)
}
}

