% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitely.R
\name{definitely}
\alias{definitely}
\title{Keeps trying a functiont until it succeeds}
\usage{
definitely(func, n_tries = 10, sleep = 1, ...)
}
\arguments{
\item{func}{A function you want to succed.}

\item{n_tries}{The number of tries before failing. 10 by default.}

\item{sleep}{Number of seconds between each try. 1 second by default.}

\item{...}{Further arguments passed to func}
}
\value{
Returns the result of func, or NULL if unsuccessful.
}
\description{
Keeps trying a functiont until it succeeds
}
\examples{
\dontrun{
Read for full example http://www.brodrigues.co/blog/2018-03-12-keep_trying/
get_data = function(){
  number = rbinom(1, 1, 0.9)
  ifelse(number == 0, "OK", stop("Error: too many calls!"))
  }
  definitely(get_data, 10, 1)
}
}
