% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Compute descriptive statistics for a data frame.}
\usage{
describe(df, ...)
}
\arguments{
\item{df}{The data frame to summarise.}

\item{...}{Optional. Columns in the data frame}
}
\value{
A data frame with descriptive statistics. If you are only interested in certain columns
you can add these columns.
}
\description{
Compute descriptive statistics for a data frame.
}
\examples{
\dontrun{
describe(dataset)
describe(dataset, col1, col2)
}
}
