% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minus_list.R
\name{\%-l\%}
\alias{\%-l\%}
\title{Substracts elements from lists of elements}
\usage{
list \%-l\% string
}
\arguments{
\item{list}{An atomic vector of elements}

\item{string}{An element (or an atomic vector of elements) to remove from list}
}
\value{
An atomic vector of elements
}
\description{
Substracts elements from lists of elements
}
\examples{
fruits <- c("Apple", "Pineapple", "Lettuce", "Kiwi", "Artichoke")
fruits \%-l\% c("Lettuce", "Artichoke")
}
