% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni.R
\name{\%ni\%}
\alias{\%ni\%}
\title{«Not in» operator. Taken from a stackoverflow answer by user `baptiste`}
\usage{
... \%ni\% NA
}
\arguments{
\item{x}{An object}

\item{y}{A list of objects}
}
\value{
True or False
}
\description{
«Not in» operator. Taken from a stackoverflow answer by user `baptiste`
}
\examples{
3 \%ni\% seq(5, 10)
}
