% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_filter.R
\name{map_filter}
\alias{map_filter}
\title{Filter a list of conditions}
\usage{
map_filter(dataset, variable, list_values, rbind = FALSE)
}
\arguments{
\item{dataset}{A `data.frame` object}

\item{variable}{The variable to filter}

\item{list_values}{An atomic vector giving the conditions we want to filter by}
}
\value{
A list containing the filtered `data.frame`s.
}
\description{
Filter a list of conditions
}
\examples{
data(mtcars)
map_filter(mtcars, "cyl", c(4, 6))
}
