% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_row.R
\name{one_row}
\alias{one_row}
\title{Keep only one row per individual}
\usage{
one_row(dataframe, id_var)
}
\arguments{
\item{dataframe}{A `data.frame` object}

\item{id_var}{A character object specifying the variable identifying individuals}
}
\value{
A `data.frame` object with only one line per individual
}
\description{
Keep only one row per individual
}
\examples{
dataframe <- as.data.frame(list("id" = c(rep(1, 3), rep(2, 4)), "x1" = rnorm(7)))
new_df <- one_row(dataframe, "id")
}

