% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_clean.R
\name{read_excel_clean}
\alias{read_excel_clean}
\title{Reads in and cleans the names of the columns of Excel files}
\usage{
read_excel_clean(path_to_excel, ...)
}
\arguments{
\item{path_to_excel}{Path to an Excel file}

\item{...}{Further arguments passed to readxl::read_excel()}
}
\value{
A data.frame containing the read Excel file
}
\description{
This is an helper function that simply reads in an Excel file and cleans the names
  of the columns in one step. The user does not need to run this by itself, but it is called by
  `read_xls_and_sheets()`.
}
\examples{
\dontrun{
read_excel_clean("my_excel_file.xlsx")
}
}

