% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_list.R
\name{read_list}
\alias{read_list}
\title{Reads a list of datasets}
\usage{
read_list(list_of_datasets, read_func, ...)
}
\arguments{
\item{list_of_datasets}{A list of datasets (names of datasets are strings)}

\item{read_func}{A function, the read function to use to read the data}

\item{...}{Further arguments passed to read_func}
}
\value{
Returns a list of the datasets
}
\description{
Reads a list of datasets
}
\examples{
\dontrun{
setwd("path/to/datasets/")
list_of_datasets <- list.files(pattern = "*.csv")
list_of_loaded_datasets <- read_list(list_of_datasets, read_func = read.csv)
}
}

