% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_transfer.R
\name{read_transfer}
\alias{read_transfer}
\title{Imports data from StatTransfer}
\usage{
read_transfer(dat, stsd, n = -1L, ok = TRUE, warn = TRUE,
  encoding = "unknown", skipNul = FALSE, sep = ",", ...)
}
\arguments{
\item{dat}{The file containing the data, in the .dat format}

\item{stsd}{The file containing the schema, in the .stsd format}

\item{...}{Further options passed down to methods}
}
\value{
A data frame
}
\description{
Imports data from StatTransfer
}
\examples{
\dontrun{
read_transfer("mtcars.dat", "mtcars.stsd")
}
}
