% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_workbook.R
\name{read_workbook}
\alias{read_workbook}
\title{Reads an Excel workbook file into a list object}
\usage{
read_workbook(path_to_excel)
}
\arguments{
\item{path_to_excel}{Path to an Excel file}
}
\value{
A list of data.frame objects where each data.frame object is an Excel sheet
}
\description{
This reads in an entire Excel workbook into a list of data frames. Each data frame
  is one the sheets that constitute the Excel workbook. This is the closest representation to
  an Excel workbook that is possible in R.
}
\examples{
\dontrun{
excel_workbook <- read_workbook("my_excel_file.xlsx")
}
}

