% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mode.R
\name{sample_mode}
\alias{sample_mode}
\title{Returns the mode of a vector}
\usage{
sample_mode(x, na.rm = FALSE)
}
\arguments{
\item{x}{A list of values}

\item{na.rm}{Should NAs be removed? FALSE by default}
}
\value{
Returns the mode of x, either a numeric if x is a list of numerics, or a character if
x is a list of characters
}
\description{
Returns the mode of a vector
}
\examples{
x <- c(3,3,3,4,5,NA, NA, NA, NA, NA)
sample_mode(x, na.rm = TRUE)
}
