% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_to_na.R
\name{set_cols_to_na}
\alias{set_cols_to_na}
\title{This function sets values to NA based on column names and indices}
\usage{
set_cols_to_na(df, variables, list_indices)
}
\arguments{
\item{df}{A data frame}

\item{variables}{A character vector of the variables to work on. Ideally returned from validate::variables(validator_object)}

\item{list_indices}{A list indices to set to NA.}
}
\value{
The original data frame with NAs at the right places.
}
\description{
This function is useful to set unlikely values to NA.
}
