% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_to_na.R
\name{set_one_col_to_na}
\alias{set_one_col_to_na}
\title{This function sets values in one column to NA.}
\usage{
set_one_col_to_na(df, variable, indices)
}
\arguments{
\item{df}{A data frame}

\item{variable}{A character vector of the variable to work on.}

\item{indices}{An atomic vector of indices to set to NA.}
}
\value{
The original data frame with NAs at the right places.
}
\description{
This function is useful to set unlikely values to NA.
}
