% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_map.R
\name{to_map}
\alias{to_map}
\title{Generalizes functions to work on lists of data}
\usage{
to_map(func)
}
\arguments{
\item{func}{A function}
}
\value{
A function that works on lists of data
}
\description{
Generalizes functions to work on lists of data
}
\examples{
\dontrun{
setwd("path/to/datasets/")
list_of_datasets <- list.files(pattern = "*.csv")
list_of_loaded_datasets <- read_list(list_of_datasets, read_func = read.csv)
# To get the summary statistics of all variables for all the datasets in the list:
summarymap <- to_map(summary)
summarymap(list_of_loaded_datasets)
# To get the summary statistics of a single variable for all the datasets in the list:
summarymap(list_of_loaded_datasets, "variable")
# This is equivalent to
purrr::map(list_of_loaded_datasets, (function(x) summary(x$variable)))
}
}
