% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_name.R
\name{invert_name}
\alias{invert_name}
\title{Inverts a surname name to name surname (or vice versa)}
\usage{
invert_name(name, to_title = F)
}
\arguments{
\item{name}{The name to invert}

\item{to_title}{Should the result be converted to title? (John Doe instead of john doe)}
}
\value{
A string with the surname and name inverted
}
\description{
Inverts a surname name to name surname (or vice versa)
}
\examples{
name <- "RODRIGUES BRUNO"
invert_name(name)
invert_name(name, to_title = T)
}
